Move loadPlugins to the separate package
